Checksum calculation

+++ src/com/jmace/MaceDrone/msp/MultiWiiClient.java
@@ -143,15 +143,15 @@ public class MultiWiiClient {
         }
         
         int length = inStream.read();
-        byte computedChecksum = (byte) inStream.read(); //command
+        int computedChecksum = inStream.read(); //command
         byte[] response = new byte[length];
         
         for (int i = 0; i < length && inStream.available() > 0; i++) {
         	response[i] = (byte) inStream.read();
-        	computedChecksum ^= response[i];
+        	computedChecksum ^= (int) response[i];
         }
         
-        byte checksum = (byte) inStream.read();
+        int checksum = inStream.read();
         
         if (computedChecksum != checksum) {
         	return new byte[0];
